% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{workshop7-package}
\alias{workshop7}
\alias{workshop7-package}
\title{workshop7: SRA 2021 Workshop 7}
\description{
Analysis package for SRA 2021 Workshop 7: Monte Carlo simulation and probability bounds analysis in R with hardly any data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jabenninghoff.github.io/workshop7/}
  \item \url{https://github.com/jabenninghoff/workshop7}
  \item Report bugs at \url{https://github.com/jabenninghoff/workshop7/issues}
}

}
\author{
\strong{Maintainer}: John Benninghoff \email{jbenninghoff@mac.com} (\href{https://orcid.org/0000-0002-6230-4742}{ORCID})

Authors:
\itemize{
  \item John Benninghoff \email{jbenninghoff@mac.com} (\href{https://orcid.org/0000-0002-6230-4742}{ORCID})
}

}
\keyword{internal}
